feat: config storybook
