+++ .storybook/style.css
@@ -1,0 +1,3 @@
+body {
+    display: contents !important;
+}
