+++ src/shared/components/button/index.tsx
@@ -1,11 +1,11 @@
 
 
-import { ForwardRefExoticComponent, HTMLAttributes, RefAttributes, forwardRef } from "react";
-import { IonButton, RouterDirection, RouterOptions, AnimationBuilder } from "@ionic/react";
 import type { JSX } from '@ionic/core/components';
+import { HTMLAttributes, RefAttributes, forwardRef } from "react";
+import { IonButton, RouterDirection, RouterOptions, AnimationBuilder } from "@ionic/react";
 import { IonicReactProps } from "@ionic/react/dist/types/components/IonicReactProps";
+
 import "./style.css";
-import React from "react";
 
 export type IButtonProps = Omit<JSX.IonButton, "routerDirection"> & Omit<HTMLAttributes<HTMLIonButtonElement>, "style"> & IonicReactProps & RefAttributes<HTMLIonButtonElement> & {
         routerLink?: string | undefined;
